+++ src/test/java/com/project/doongdoong/domain/counsel/service/CounselStatisticsServiceImplTest.java
@@ -1,5 +1,6 @@
 package com.project.doongdoong.domain.counsel.service;
 
+import com.project.doongdoong.domain.counsel.application.port.in.CounselStatisticsService;
 import com.project.doongdoong.domain.counsel.dto.CounselRankList;
 import com.project.doongdoong.domain.counsel.model.CounselCacheKey;
 import com.project.doongdoong.domain.counsel.model.CounselRank;
