+++ src/test/java/com/project/doongdoong/module/ControllerTestSupport.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.doongdoong.domain.analysis.adapter.in.web.AnalysisController;
 import com.project.doongdoong.domain.analysis.application.port.in.AnalysisService;
 import com.project.doongdoong.domain.answer.application.port.in.AnswerService;
-import com.project.doongdoong.domain.counsel.controller.CounselController;
-import com.project.doongdoong.domain.counsel.service.CounselService;
-import com.project.doongdoong.domain.counsel.service.CounselStatisticsService;
+import com.project.doongdoong.domain.counsel.adapter.in.web.CounselController;
+import com.project.doongdoong.domain.counsel.application.port.in.CounselService;
+import com.project.doongdoong.domain.counsel.application.port.in.CounselStatisticsService;
 import com.project.doongdoong.domain.user.controller.UserController;
 import com.project.doongdoong.domain.user.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
